[Chore] Split config.rs into schema, load, env, validate and tuning
`config.rs` was 3 044 lines carrying five unrelated jobs: the shape of
every setting, reading a file or `opendal://` source, `${env:}` expansion
and `HARUKI_*` overrides, the rules a config must satisfy, and deriving
pool widths from the host. Nothing separated them, so any one of them was
read against the whole file.

This is a pure move. Every item keeps its body byte-for-byte; an
item-by-item comparison against the original confirms 147 items in and
147 items out, none lost, none added. `config.rs` stays the facade and
re-exports the schema, so every external path -- `crate::core::config::
AppConfig` and the rest -- is unchanged.

Two helpers moved to where they belong rather than where they sat:
`dedupe_*_formats` normalise a configured list rather than rejecting
anything, so they follow the shape into `schema` instead of `validate`.

Visibility that meant nothing inside one file now carries the structure:
the edges are `schema` (leaf) <- `validate`, `tuning`, `env` <- `load`,
which is a DAG with loading on top.

Tests stay in the facade for this commit so the move can be read as a
move; splitting them by domain is a separate change.
