+++ src/core/config.rs
@@ -1,2044 +1,28 @@
-use std::collections::BTreeMap;
-use std::env;
-use std::fs;
-use std::path::{Path, PathBuf};
-use std::str::FromStr;
-
-use serde::de::{self, Deserializer};
-use serde::{Deserialize, Serialize};
-use yaml_serde::{Mapping, Value};
-
-use crate::core::errors::ConfigError;
-
-const CONFIG_URI_ENV: &str = "HARUKI_CONFIG_URI";
-const CONFIG_OPENDAL_SCHEME_ENV: &str = "HARUKI_CONFIG_OPENDAL_SCHEME";
-const CONFIG_OPENDAL_ROOT_ENV: &str = "HARUKI_CONFIG_OPENDAL_ROOT";
-const CONFIG_OPENDAL_OPTION_PREFIX: &str = "HARUKI_CONFIG_OPENDAL_OPTION_";
-const CONFIG_OPENDAL_URI_PREFIX: &str = "opendal://";
-pub const CURRENT_CONFIG_VERSION: u32 = 3;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ConfigStorageUri {
-    provider: String,
-    path: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct AppConfig {
-    pub config_version: u32,
-    pub server: ServerConfig,
-    pub logging: LoggingConfig,
-    pub execution: ExecutionConfig,
-    pub backends: BackendsConfig,
-    pub resources: ResourcesConfig,
-    pub concurrency: ConcurrencyConfig,
-    pub storage: StorageConfig,
-    pub git_sync: GitSyncConfig,
-    pub regions: BTreeMap<String, RegionConfig>,
-}
-
-impl Default for AppConfig {
-    fn default() -> Self {
-        Self {
-            config_version: CURRENT_CONFIG_VERSION,
-            server: ServerConfig::default(),
-            logging: LoggingConfig::default(),
-            execution: ExecutionConfig::default(),
-            backends: BackendsConfig::default(),
-            resources: ResourcesConfig::default(),
-            concurrency: ConcurrencyConfig::default(),
-            storage: StorageConfig::default(),
-            git_sync: GitSyncConfig::default(),
-            regions: BTreeMap::new(),
-        }
-    }
-}
-
-impl AppConfig {
-    pub async fn load_default() -> Result<Self, ConfigError> {
-        if let Some(uri) = env::var(CONFIG_URI_ENV)
-            .ok()
-            .map(|uri| uri.trim().to_string())
-            .filter(|uri| !uri.is_empty())
-        {
-            return Self::load_from_opendal_uri(&uri).await;
-        }
-
-        let candidates = candidate_paths();
-        for candidate in &candidates {
-            if candidate.exists() {
-                return Self::load_from_path(candidate);
-            }
-        }
-
-        Err(ConfigError::MissingConfigFile(
-            candidates
-                .iter()
-                .map(|path| path.display().to_string())
-                .collect::<Vec<_>>()
-                .join(", "),
-        ))
-    }
-
-    pub fn load_from_path(path: impl AsRef<Path>) -> Result<Self, ConfigError> {
-        let path = path.as_ref().to_path_buf();
-        let raw = fs::read_to_string(&path).map_err(|source| ConfigError::Read {
-            path: path.clone(),
-            source,
-        })?;
-        Self::load_from_str(path, &raw)
-    }
-
-    pub async fn load_from_opendal_uri(uri: &str) -> Result<Self, ConfigError> {
-        let storage_uri = parse_config_storage_uri(uri)?;
-        let (scheme, options) = config_storage_provider_options()?;
-
-        opendal::init_default_registry();
-        let operator = opendal::Operator::via_iter(&scheme, options).map_err(|source| {
-            ConfigError::ConfigStorageProvider {
-                provider: storage_uri.provider.clone(),
-                source,
-            }
-        })?;
-        let bytes = operator.read(&storage_uri.path).await.map_err(|source| {
-            ConfigError::ConfigStorageRead {
-                uri: uri.to_string(),
-                source,
-            }
-        })?;
-        let raw = String::from_utf8(bytes.to_vec()).map_err(|source| ConfigError::InvalidUtf8 {
-            path: uri.to_string(),
-            source,
-        })?;
-
-        Self::load_from_str(PathBuf::from(uri), &raw)
-    }
-
-    fn load_from_str(path: PathBuf, raw: &str) -> Result<Self, ConfigError> {
-        let mut value: Value = yaml_serde::from_str(raw).map_err(|source| ConfigError::Parse {
-            path: path.clone(),
-            source,
-        })?;
-        expand_env_references(&mut value)?;
-        apply_env_overrides(&mut value)?;
-
-        let mut config: Self =
-            yaml_serde::from_value(value).map_err(|source| ConfigError::Parse {
-                path: path.clone(),
-                source,
-            })?;
-        config.resolve_env_overrides()?;
-        config.validate()?;
-        Ok(config)
-    }
-
-    pub fn validate(&self) -> Result<(), ConfigError> {
-        if self.config_version != CURRENT_CONFIG_VERSION {
-            return Err(ConfigError::UnsupportedVersion(self.config_version));
-        }
-        validate_region_names(self)?;
-        validate_runtime_settings(self)?;
-        validate_auth_config(&self.server.auth)?;
-        validate_regions(self)?;
-        validate_asset_studio_read_kinds(&self.backends.asset_studio.read_kinds)?;
-        warn_media_fallback_backend_options(&self.backends.media);
-
-        Ok(())
-    }
-
-    pub fn effective_concurrency(&self) -> ConcurrencyConfig {
-        self.effective_concurrency_for_cpus(available_cpu_count())
-    }
-
-    pub fn effective_concurrency_for_cpus(&self, cpus: usize) -> ConcurrencyConfig {
-        self.concurrency.effective_for_cpus_with_budget(
-            cpus,
-            self.resources.cpu.effective_budget_for_cpus(cpus),
-        )
-    }
-
-    pub fn effective_cpu_budget(&self) -> usize {
-        self.resources.cpu.effective_budget()
-    }
-
-    pub fn enabled_regions(&self) -> Vec<String> {
-        self.regions
-            .iter()
-            .filter_map(|(name, region)| region.enabled.then_some(name.clone()))
-            .collect()
-    }
-
-    fn resolve_env_overrides(&mut self) -> Result<(), ConfigError> {
-        resolve_backend_env_overrides(self)?;
-        resolve_concurrency_env_overrides(self)?;
-        resolve_resource_env_overrides(self)?;
-        resolve_config_secret_env_overrides(self)
-    }
-}
-
-fn resolve_backend_env_overrides(config: &mut AppConfig) -> Result<(), ConfigError> {
-    if let Ok(value) = env::var("HARUKI_MEDIA_BACKEND") {
-        config.backends.media.backend = value.parse()?;
-    }
-    if let Ok(value) = env::var("HARUKI_ASSET_STUDIO_READ_BATCH_SIZE") {
-        config.backends.asset_studio.read_batch_size =
-            parse_positive_usize("backends.asset_studio.read_batch_size", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_ASSET_STUDIO_IMAGE_FORMAT") {
-        config.backends.asset_studio.image_format =
-            non_empty_option(normalize_asset_studio_image_format(&value)?);
-    }
-    if let Ok(value) = env::var("HARUKI_ASSET_HTTP_VERSION") {
-        config.server.asset_http_version = value.parse()?;
-    }
-    Ok(())
-}
-
-fn resolve_concurrency_env_overrides(config: &mut AppConfig) -> Result<(), ConfigError> {
-    if let Ok(value) = env::var("HARUKI_MEDIA_ENCODE_CONCURRENCY") {
-        let parsed = parse_positive_usize("concurrency.media_encode", &value)?;
-        config.concurrency.media_encode = parsed;
-        config.concurrency.audio_encode = parsed;
-        config.concurrency.video_encode = parsed;
-    }
-    if let Ok(value) = env::var("HARUKI_AUDIO_ENCODE_CONCURRENCY") {
-        config.concurrency.audio_encode = parse_positive_usize("concurrency.audio_encode", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_VIDEO_ENCODE_CONCURRENCY") {
-        config.concurrency.video_encode = parse_positive_usize("concurrency.video_encode", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_DOWNLOAD_CONCURRENCY") {
-        config.concurrency.download = parse_positive_usize("concurrency.download", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_POST_PROCESS_CONCURRENCY") {
-        config.concurrency.post_process = parse_positive_usize("concurrency.post_process", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_CONCURRENCY_AUTO_TUNE") {
-        config.concurrency.auto_tune = parse_bool_env("concurrency.auto_tune", &value)?;
-    }
-    Ok(())
-}
-
-fn resolve_resource_env_overrides(config: &mut AppConfig) -> Result<(), ConfigError> {
-    if let Ok(value) = env::var("HARUKI_CPU_BUDGET_AUTO") {
-        config.resources.cpu.budget_auto = parse_bool_env("resources.cpu.budget_auto", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_CPU_BUDGET_RATIO") {
-        config.resources.cpu.budget_ratio =
-            parse_cpu_ratio_env("resources.cpu.budget_ratio", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_CPU_RESERVED") {
-        config.resources.cpu.reserved = parse_usize_env("resources.cpu.reserved", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_CPU_THROTTLE_ENABLED") {
-        config.resources.cpu.throttle.enabled =
-            parse_bool_env("resources.cpu.throttle.enabled", &value)?;
-    }
-    if let Ok(value) = env::var("HARUKI_CPU_THROTTLE_SAMPLE_MS") {
-        config.resources.cpu.throttle.sample_ms =
-            parse_positive_usize("resources.cpu.throttle.sample_ms", &value)? as u64;
-    }
-    if let Ok(value) = env::var("HARUKI_MAX_IN_FLIGHT_BUNDLE_BYTES") {
-        config.resources.memory.max_in_flight_bundle_bytes =
-            parse_usize_env("resources.memory.max_in_flight_bundle_bytes", &value)?;
-    }
-    Ok(())
-}
-
-fn resolve_config_secret_env_overrides(config: &mut AppConfig) -> Result<(), ConfigError> {
-    resolve_secret_env(
-        "git_sync.chart_hashes.password",
-        &mut config.git_sync.chart_hashes.password,
-    )?;
-    for (idx, provider) in config.storage.providers.iter_mut().enumerate() {
-        resolve_secret_env(
-            &format!("storage.providers[{idx}].access_key"),
-            &mut provider.access_key,
-        )?;
-        resolve_secret_env(
-            &format!("storage.providers[{idx}].secret_key"),
-            &mut provider.secret_key,
-        )?;
-    }
-    for (region_name, region) in &mut config.regions {
-        resolve_secret_env(
-            &format!("regions.{region_name}.crypto.aes_key_hex"),
-            &mut region.crypto.aes_key_hex,
-        )?;
-        resolve_secret_env(
-            &format!("regions.{region_name}.crypto.aes_iv_hex"),
-            &mut region.crypto.aes_iv_hex,
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_region_names(config: &AppConfig) -> Result<(), ConfigError> {
-    for region_name in config.regions.keys() {
-        if region_name.to_lowercase() != *region_name {
-            return Err(ConfigError::InvalidRegionName(region_name.clone()));
-        }
-    }
-    Ok(())
-}
-
-fn validate_runtime_settings(config: &AppConfig) -> Result<(), ConfigError> {
-    let budget_ratio = config.resources.cpu.budget_ratio;
-    if !(0.0..=1.0).contains(&budget_ratio) || budget_ratio == 0.0 {
-        return Err(ConfigError::InvalidValue {
-            field: "resources.cpu.budget_ratio".to_string(),
-            value: budget_ratio.to_string(),
-            expected: "a number greater than 0 and less than or equal to 1".to_string(),
-        });
-    }
-    validate_positive_setting(
-        "backends.asset_studio.read_batch_size",
-        config.backends.asset_studio.read_batch_size,
-    )?;
-    validate_positive_setting("concurrency.media_encode", config.concurrency.media_encode)?;
-    validate_positive_setting("concurrency.audio_encode", config.concurrency.audio_encode)?;
-    validate_positive_setting("concurrency.video_encode", config.concurrency.video_encode)?;
-    if let Some(image_format) = &config.backends.asset_studio.image_format {
-        validate_asset_studio_image_format(image_format)?;
-    }
-    validate_image_backend(&config.backends.image)
-}
-
-fn validate_positive_setting(field: &str, value: usize) -> Result<(), ConfigError> {
-    if value == 0 {
-        return Err(ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: "0".to_string(),
-            expected: "a positive integer".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn validate_auth_config(auth: &AuthConfig) -> Result<(), ConfigError> {
-    if !auth.enabled {
-        return Ok(());
-    }
-    let has_credential = auth
-        .bearer_token
-        .as_deref()
-        .is_some_and(|token| !token.is_empty())
-        || auth
-            .user_agent_prefix
-            .as_deref()
-            .is_some_and(|prefix| !prefix.is_empty());
-    if has_credential {
-        return Ok(());
-    }
-    Err(ConfigError::InvalidValue {
-        field: "server.auth".to_string(),
-        value: "enabled=true with no credentials".to_string(),
-        expected: "a non-empty bearer_token or user_agent_prefix when auth is enabled".to_string(),
-    })
-}
-
-fn validate_regions(config: &AppConfig) -> Result<(), ConfigError> {
-    for (region_name, region) in &config.regions {
-        validate_image_export_config(region_name, &region.export.images)?;
-        validate_video_export_config(region_name, &region.export.video)?;
-        validate_audio_export_config(region_name, &region.export.audio)?;
-        validate_haruki_3d_export_config(region_name, &region.export.haruki_3d)?;
-        if region.enabled {
-            validate_region_crypto(region_name, &region.crypto)?;
-            validate_region_filter_regexes(region_name, region)?;
-        }
-    }
-    Ok(())
-}
-
-fn resolve_secret_env(field: &str, value: &mut Option<String>) -> Result<(), ConfigError> {
-    let Some(raw) = value.as_deref().map(str::trim) else {
-        return Ok(());
-    };
-
-    let Some(name) = raw
-        .strip_prefix("${env:")
-        .and_then(|rest| rest.strip_suffix('}'))
-        .map(str::trim)
-    else {
-        return Ok(());
-    };
-
-    let resolved = env::var(name).map_err(|_| ConfigError::MissingEnvironmentVariable {
-        field: field.to_string(),
-        name: name.to_string(),
-    })?;
-    *value = Some(resolved);
-    Ok(())
-}
-
-fn parse_config_storage_uri(uri: &str) -> Result<ConfigStorageUri, ConfigError> {
-    let Some(raw) = uri.strip_prefix(CONFIG_OPENDAL_URI_PREFIX) else {
-        return Err(ConfigError::InvalidConfigUri {
-            uri: uri.to_string(),
-            reason:
-                "only opendal:// config URIs are supported; use HARUKI_CONFIG_PATH for local files"
-                    .to_string(),
-        });
-    };
-
-    let raw = raw.trim_start_matches('/');
-    let Some((provider, path)) = raw.split_once('/') else {
-        return Err(ConfigError::InvalidConfigUri {
-            uri: uri.to_string(),
-            reason: "expected opendal://<provider>/<path>".to_string(),
-        });
-    };
-    let provider = provider.trim();
-    let path = path.trim().trim_matches('/').replace('\\', "/");
-
-    if provider.is_empty() {
-        return Err(ConfigError::InvalidConfigUri {
-            uri: uri.to_string(),
-            reason: "provider is empty".to_string(),
-        });
-    }
-    if path.is_empty() {
-        return Err(ConfigError::InvalidConfigUri {
-            uri: uri.to_string(),
-            reason: "path is empty".to_string(),
-        });
-    }
-
-    Ok(ConfigStorageUri {
-        provider: provider.to_string(),
-        path,
-    })
-}
-
-fn config_storage_provider_options() -> Result<(String, BTreeMap<String, String>), ConfigError> {
-    let scheme = env::var(CONFIG_OPENDAL_SCHEME_ENV)
-        .ok()
-        .map(|scheme| scheme.trim().to_ascii_lowercase())
-        .filter(|scheme| !scheme.is_empty())
-        .ok_or_else(|| ConfigError::MissingEnvironmentVariable {
-            field: CONFIG_URI_ENV.to_string(),
-            name: CONFIG_OPENDAL_SCHEME_ENV.to_string(),
-        })?;
-
-    let mut options = BTreeMap::new();
-    if let Some(root) = env::var(CONFIG_OPENDAL_ROOT_ENV)
-        .ok()
-        .map(|root| root.trim().to_string())
-        .filter(|root| !root.is_empty())
-    {
-        options.insert("root".to_string(), root);
-    }
-
-    for (name, value) in env::vars().filter(|(name, value)| {
-        name.starts_with(CONFIG_OPENDAL_OPTION_PREFIX)
-            && name.len() > CONFIG_OPENDAL_OPTION_PREFIX.len()
-            && !value.trim().is_empty()
-    }) {
-        let key = name
-            .strip_prefix(CONFIG_OPENDAL_OPTION_PREFIX)
-            .expect("prefix was checked")
-            .to_ascii_lowercase();
-        if key.is_empty() {
-            return Err(ConfigError::InvalidConfigBootstrap {
-                name,
-                reason: "OpenDAL option key is empty".to_string(),
-            });
-        }
-        options.insert(key, value);
-    }
-
-    Ok((scheme, options))
-}
-
-fn expand_env_references(value: &mut Value) -> Result<(), ConfigError> {
-    match value {
-        Value::String(raw) => {
-            if let Some(expanded) = expand_env_references_in_string(raw)? {
-                *raw = expanded;
-            }
-        }
-        Value::Sequence(items) => {
-            for item in items {
-                expand_env_references(item)?;
-            }
-        }
-        Value::Mapping(map) => {
-            for (_, value) in map.iter_mut() {
-                expand_env_references(value)?;
-            }
-        }
-        Value::Tagged(tagged) => expand_env_references(&mut tagged.value)?,
-        Value::Null | Value::Bool(_) | Value::Number(_) => {}
-    }
-
-    Ok(())
-}
-
-fn expand_env_references_in_string(raw: &str) -> Result<Option<String>, ConfigError> {
-    let Some(mut start) = raw.find("${env:") else {
-        return Ok(None);
-    };
-
-    let mut expanded = String::with_capacity(raw.len());
-    let mut cursor = 0;
-
-    while start < raw.len() {
-        expanded.push_str(&raw[cursor..start]);
-        let name_start = start + "${env:".len();
-        let Some(relative_end) = raw[name_start..].find('}') else {
-            // An unclosed `${env:` is almost always a typo (e.g. a missing `}` on a secret field).
-            // Failing loudly beats silently treating it as a literal value, which would surface
-            // later as a confusing hex/decrypt error.
-            return Err(ConfigError::InvalidValue {
-                field: "config file".to_string(),
-                value: "${env:...".to_string(),
-                expected: "a closed ${env:VAR} reference (missing closing '}')".to_string(),
-            });
-        };
-        let end = name_start + relative_end;
-        let name = raw[name_start..end].trim();
-        let value = env::var(name).map_err(|_| ConfigError::MissingEnvironmentVariable {
-            field: "config file".to_string(),
-            name: name.to_string(),
-        })?;
-        expanded.push_str(&value);
-        cursor = end + 1;
-
-        let Some(next) = raw[cursor..].find("${env:") else {
-            break;
-        };
-        start = cursor + next;
-    }
-
-    expanded.push_str(&raw[cursor..]);
-    Ok(Some(expanded))
-}
-
-fn apply_env_overrides(root: &mut Value) -> Result<(), ConfigError> {
-    let overrides = env::vars()
-        .filter(|(name, _)| name.starts_with("HARUKI__"))
-        .collect::<BTreeMap<_, _>>();
-
-    for (name, raw_value) in overrides {
-        let path = parse_env_override_path(&name)?;
-        let value = parse_env_override_value(&raw_value);
-        apply_env_override(root, &name, &path, value)?;
-    }
-
-    Ok(())
-}
-
-fn parse_env_override_path(name: &str) -> Result<Vec<String>, ConfigError> {
-    let raw_path =
-        name.strip_prefix("HARUKI__")
-            .ok_or_else(|| ConfigError::InvalidConfigBootstrap {
-                name: name.to_string(),
-                reason: "override names must start with HARUKI__".to_string(),
-            })?;
-
-    if raw_path.is_empty() {
-        return Err(ConfigError::InvalidConfigBootstrap {
-            name: name.to_string(),
-            reason: "override path is empty".to_string(),
-        });
-    }
-
-    raw_path
-        .split("__")
-        .map(|segment| {
-            if segment.is_empty() {
-                Err(ConfigError::InvalidConfigBootstrap {
-                    name: name.to_string(),
-                    reason: "override path contains an empty segment".to_string(),
-                })
-            } else {
-                Ok(segment.to_ascii_lowercase())
-            }
-        })
-        .collect()
-}
-
-fn parse_env_override_value(raw: &str) -> Value {
-    if raw.is_empty() {
-        return Value::String(String::new());
-    }
-
-    yaml_serde::from_str::<Value>(raw).unwrap_or_else(|_| Value::String(raw.to_string()))
-}
-
-fn apply_env_override(
-    root: &mut Value,
-    name: &str,
-    path: &[String],
-    value: Value,
-) -> Result<(), ConfigError> {
-    if path.is_empty() {
-        return Err(ConfigError::InvalidConfigBootstrap {
-            name: name.to_string(),
-            reason: "override path is empty".to_string(),
-        });
-    }
-
-    let mut current = root;
-    for (idx, segment) in path.iter().enumerate() {
-        let is_last = idx + 1 == path.len();
-        if is_last {
-            set_env_override_leaf(current, segment, value);
-            return Ok(());
-        }
-
-        current =
-            descend_env_override_path(current, segment, path.get(idx + 1).map(String::as_str));
-    }
-
-    Ok(())
-}
-
-fn set_env_override_leaf(current: &mut Value, segment: &str, value: Value) {
-    if let Ok(index) = segment.parse::<usize>() {
-        match current {
-            Value::Sequence(items) => {
-                if items.len() <= index {
-                    items.resize(index + 1, Value::Null);
-                }
-                items[index] = value;
-                return;
-            }
-            Value::Null => {
-                let mut items = Vec::new();
-                items.resize(index + 1, Value::Null);
-                items[index] = value;
-                *current = Value::Sequence(items);
-                return;
-            }
-            _ => {}
-        }
-    }
-
-    if !matches!(current, Value::Mapping(_)) {
-        *current = Value::Mapping(Mapping::new());
-    }
-
-    if let Value::Mapping(map) = current {
-        map.insert(Value::String(segment.to_string()), value);
-    }
-}
-
-fn descend_env_override_path<'a>(
-    current: &'a mut Value,
-    segment: &str,
-    next_segment: Option<&str>,
-) -> &'a mut Value {
-    let next_is_index = next_segment.is_some_and(|next| next.parse::<usize>().is_ok());
-
-    if let Ok(index) = segment.parse::<usize>() {
-        if matches!(current, Value::Null) {
-            *current = Value::Sequence(Vec::new());
-        }
-        if let Value::Sequence(items) = current {
-            if items.len() <= index {
-                items.resize_with(index + 1, Value::default);
-            }
-            if matches!(items[index], Value::Null) {
-                items[index] = env_override_default_child(next_is_index);
-            }
-            return &mut items[index];
-        }
-    }
-
-    if !matches!(current, Value::Mapping(_)) {
-        *current = Value::Mapping(Mapping::new());
-    }
-
-    if let Value::Mapping(map) = current {
-        return map
-            .entry(Value::String(segment.to_string()))
-            .or_insert_with(|| env_override_default_child(next_is_index));
-    }
-
-    unreachable!("current value was normalized into a mapping")
-}
-
-fn env_override_default_child(sequence: bool) -> Value {
-    if sequence {
-        Value::Sequence(Vec::new())
-    } else {
-        Value::Mapping(Mapping::new())
-    }
-}
-
-fn non_empty_option(value: String) -> Option<String> {
-    if value.trim().is_empty() {
-        None
-    } else {
-        Some(value)
-    }
-}
-
-fn parse_positive_usize(field: &str, value: &str) -> Result<usize, ConfigError> {
-    let trimmed = value.trim();
-    let parsed = trimmed
-        .parse::<usize>()
-        .map_err(|_| ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: trimmed.to_string(),
-            expected: "a positive integer".to_string(),
-        })?;
-    if parsed == 0 {
-        Err(ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: trimmed.to_string(),
-            expected: "a positive integer".to_string(),
-        })
-    } else {
-        Ok(parsed)
-    }
-}
-
-fn parse_usize_env(field: &str, value: &str) -> Result<usize, ConfigError> {
-    let trimmed = value.trim();
-    trimmed
-        .parse::<usize>()
-        .map_err(|_| ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: trimmed.to_string(),
-            expected: "a non-negative integer".to_string(),
-        })
-}
-
-fn parse_cpu_ratio_env(field: &str, value: &str) -> Result<f64, ConfigError> {
-    let trimmed = value.trim();
-    trimmed
-        .parse::<f64>()
-        .map_err(|_| ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: trimmed.to_string(),
-            expected: "a number greater than 0 and less than or equal to 1".to_string(),
-        })
-}
-
-fn validate_region_crypto(region_name: &str, crypto: &CryptoConfig) -> Result<(), ConfigError> {
-    if let Some(key_hex) = &crypto.aes_key_hex {
-        validate_aes_hex(
-            &format!("regions.{region_name}.crypto.aes_key_hex"),
-            key_hex,
-            &[16, 24, 32],
-        )?;
-    }
-    if let Some(iv_hex) = &crypto.aes_iv_hex {
-        validate_aes_hex(
-            &format!("regions.{region_name}.crypto.aes_iv_hex"),
-            iv_hex,
-            &[16],
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_aes_hex(
-    field: &str,
-    value: &str,
-    allowed_lengths: &[usize],
-) -> Result<(), ConfigError> {
-    let bytes = hex::decode(value).map_err(|_| ConfigError::InvalidValue {
-        field: field.to_string(),
-        value: "<redacted>".to_string(),
-        expected: "a valid hexadecimal string".to_string(),
-    })?;
-    if !allowed_lengths.contains(&bytes.len()) {
-        return Err(ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: format!("{} byte(s)", bytes.len()),
-            expected: format!("hex decoding to one of {allowed_lengths:?} bytes"),
-        });
-    }
-    Ok(())
-}
-
-fn validate_region_filter_regexes(
-    region_name: &str,
-    region: &RegionConfig,
-) -> Result<(), ConfigError> {
-    let filters = &region.filters;
-    validate_regex_patterns(
-        &format!("regions.{region_name}.filters.start_app"),
-        &filters.start_app,
-    )?;
-    validate_regex_patterns(
-        &format!("regions.{region_name}.filters.on_demand"),
-        &filters.on_demand,
-    )?;
-    validate_regex_patterns(
-        &format!("regions.{region_name}.filters.skip"),
-        &filters.skip,
-    )?;
-    validate_regex_patterns(
-        &format!("regions.{region_name}.filters.priority"),
-        &filters.priority,
-    )?;
-    if let Some(raw_bundles) = &region.export.raw_bundles {
-        validate_regex_patterns(
-            &format!("regions.{region_name}.export.raw_bundles.include"),
-            &raw_bundles.include,
-        )?;
-        validate_regex_patterns(
-            &format!("regions.{region_name}.export.raw_bundles.exclude"),
-            &raw_bundles.exclude,
-        )?;
-    }
-    Ok(())
-}
-
-fn validate_regex_patterns(field: &str, patterns: &[String]) -> Result<(), ConfigError> {
-    for pattern in patterns {
-        regex::Regex::new(pattern).map_err(|source| ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: pattern.clone(),
-            expected: format!("a valid regular expression ({source})"),
-        })?;
-    }
-    Ok(())
-}
-
-fn normalize_asset_studio_image_format(value: &str) -> Result<String, ConfigError> {
-    let normalized = value.trim().to_lowercase();
-    validate_asset_studio_image_format(&normalized)?;
-    Ok(normalized)
-}
-
-fn validate_asset_studio_image_format(value: &str) -> Result<(), ConfigError> {
-    match value.trim().to_lowercase().as_str() {
-        "raw_rgba" => Ok(()),
-        other => Err(ConfigError::InvalidValue {
-            field: "backends.asset_studio.image_format".to_string(),
-            value: other.to_string(),
-            expected: "raw_rgba".to_string(),
-        }),
-    }
-}
-
-fn validate_image_backend(image: &ImageBackendConfig) -> Result<(), ConfigError> {
-    match image.backend {
-        ImageBackend::Rust => {}
-    }
-    if !(1..=100).contains(&image.jpeg_quality) {
-        return Err(ConfigError::InvalidValue {
-            field: "backends.image.jpeg_quality".to_string(),
-            value: image.jpeg_quality.to_string(),
-            expected: "an integer from 1 to 100".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn validate_image_export_config(
-    region_name: &str,
-    images: &ImageExportConfig,
-) -> Result<(), ConfigError> {
-    let formats = images.output_formats();
-    if formats.is_empty() {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.images.formats"),
-            value: "[]".to_string(),
-            expected: "at least one of png, jpg, or webp".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn validate_video_export_config(
-    region_name: &str,
-    video: &VideoExportConfig,
-) -> Result<(), ConfigError> {
-    let formats = video.output_formats();
-    if formats.is_empty() {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.video.formats"),
-            value: "[]".to_string(),
-            expected: "at least one of m2v or mp4".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn validate_audio_export_config(
-    region_name: &str,
-    audio: &AudioExportConfig,
-) -> Result<(), ConfigError> {
-    let formats = audio.output_formats();
-    if formats.is_empty() {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.audio.formats"),
-            value: "[]".to_string(),
-            expected: "at least one of wav, flac, or mp3".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn validate_haruki_3d_export_config(
-    region_name: &str,
-    haruki_3d: &Haruki3dExportConfig,
-) -> Result<(), ConfigError> {
-    if !haruki_3d.enabled {
-        return Ok(());
-    }
-    for (field, value) in [
-        ("exporter_path", &haruki_3d.exporter_path),
-        ("master_dir", &haruki_3d.master_dir),
-        ("output_dir", &haruki_3d.output_dir),
-        ("manifest_file", &haruki_3d.manifest_file),
-    ] {
-        if value.trim().is_empty() {
-            return Err(ConfigError::InvalidValue {
-                field: format!("regions.{region_name}.export.haruki_3d.{field}"),
-                value: value.clone(),
-                expected: "a non-empty path".to_string(),
-            });
-        }
-    }
-    if haruki_3d.work_dir.trim().is_empty() && haruki_3d.staging_dir.trim().is_empty() {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.haruki_3d.work_dir"),
-            value: haruki_3d.work_dir.clone(),
-            expected: "a non-empty path".to_string(),
-        });
-    }
-    if haruki_3d.include.is_empty() {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.haruki_3d.include"),
-            value: "[]".to_string(),
-            expected: "at least one include pattern".to_string(),
-        });
-    }
-    if let Some(value) = haruki_3d
-        .role_character3d_ids
-        .iter()
-        .find(|value| **value <= 0)
-    {
-        return Err(ConfigError::InvalidValue {
-            field: format!("regions.{region_name}.export.haruki_3d.role_character3d_ids"),
-            value: value.to_string(),
-            expected: "positive character3d ids".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn dedupe_image_formats(formats: Vec<ImageOutputFormat>) -> Vec<ImageOutputFormat> {
-    let mut output = Vec::new();
-    for format in formats {
-        if !output.contains(&format) {
-            output.push(format);
-        }
-    }
-    output
-}
-
-fn dedupe_video_formats(formats: Vec<VideoOutputFormat>) -> Vec<VideoOutputFormat> {
-    let mut output = Vec::new();
-    for format in formats {
-        if !output.contains(&format) {
-            output.push(format);
-        }
-    }
-    output
-}
-
-fn dedupe_audio_formats(formats: Vec<AudioOutputFormat>) -> Vec<AudioOutputFormat> {
-    let mut output = Vec::new();
-    for format in formats {
-        if !output.contains(&format) {
-            output.push(format);
-        }
-    }
-    output
-}
-
-fn validate_asset_studio_read_kinds(
-    read_kinds: &BTreeMap<String, String>,
-) -> Result<(), ConfigError> {
-    for (asset_type, kind) in read_kinds {
-        if asset_type.trim().is_empty() {
-            return Err(ConfigError::InvalidValue {
-                field: "backends.asset_studio.read_kinds".to_string(),
-                value: asset_type.clone(),
-                expected: "non-empty AssetStudio type selector".to_string(),
-            });
-        }
-        validate_asset_studio_read_kind(
-            &format!("backends.asset_studio.read_kinds.{asset_type}"),
-            kind,
-        )?;
-    }
-    Ok(())
-}
-
-fn warn_media_fallback_backend_options(media: &MediaBackendConfig) {
-    match media.backend {
-        MediaBackend::Ffi => {}
-        MediaBackend::Cli => {
-            tracing::warn!("backends.media.backend=cli is a fallback mode; production Linux builds should prefer ffi")
-        }
-        MediaBackend::Auto => tracing::warn!(
-            "backends.media.backend=auto is a fallback mode; production Linux builds should prefer ffi"
-        ),
-    }
-}
-
-fn validate_asset_studio_read_kind(field: &str, value: &str) -> Result<(), ConfigError> {
-    match value.trim().to_lowercase().as_str() {
-        "auto" | "raw" | "typetree_json" | "image" | "image_archive" | "audio" | "video"
-        | "font" | "shader" | "text" | "text_bytes" | "mesh" | "obj" | "animator" | "fbx" => {
-            Ok(())
-        }
-        other => Err(ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: other.to_string(),
-            expected: "auto, raw, typetree_json, image, image_archive, audio, video, font, shader, text, text_bytes, mesh, obj, animator, or fbx".to_string(),
-        }),
-    }
-}
-
-fn parse_bool_env(field: &str, value: &str) -> Result<bool, ConfigError> {
-    let trimmed = value.trim();
-    match trimmed.to_lowercase().as_str() {
-        "1" | "true" | "yes" | "on" => Ok(true),
-        "0" | "false" | "no" | "off" => Ok(false),
-        _ => Err(ConfigError::InvalidValue {
-            field: field.to_string(),
-            value: trimmed.to_string(),
-            expected: "true or false".to_string(),
-        }),
-    }
-}
-
-fn candidate_paths() -> Vec<PathBuf> {
-    let mut candidates = Vec::new();
-    if let Ok(path) = env::var("HARUKI_CONFIG_PATH") {
-        candidates.push(PathBuf::from(path));
-    }
-    candidates.push(PathBuf::from("haruki-asset-configs.yaml"));
-    candidates.push(PathBuf::from("../haruki-asset-configs.yaml"));
-    candidates.push(PathBuf::from("../../haruki-asset-configs.yaml"));
-    candidates
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct ServerConfig {
-    pub host: String,
-    pub port: u16,
-    pub proxy: Option<String>,
-    pub asset_http_version: AssetHttpVersion,
-    pub auth: AuthConfig,
-    pub tls: TlsConfig,
-}
-
-impl Default for ServerConfig {
-    fn default() -> Self {
-        Self {
-            host: "0.0.0.0".to_string(),
-            port: 8080,
-            proxy: None,
-            asset_http_version: AssetHttpVersion::Auto,
-            auth: AuthConfig::default(),
-            tls: TlsConfig::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum AssetHttpVersion {
-    #[default]
-    Auto,
-    Http1,
-}
-
-impl FromStr for AssetHttpVersion {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "auto" => Ok(Self::Auto),
-            "http1" | "http1_only" | "http/1" | "http/1.1" => Ok(Self::Http1),
-            other => Err(ConfigError::InvalidValue {
-                field: "server.asset_http_version".to_string(),
-                value: other.to_string(),
-                expected: "auto or http1".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct AuthConfig {
-    pub enabled: bool,
-    pub user_agent_prefix: Option<String>,
-    pub bearer_token: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct TlsConfig {
-    pub enabled: bool,
-    pub cert_file: Option<String>,
-    pub key_file: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum LogFormat {
-    #[default]
-    Pretty,
-    Json,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct LoggingConfig {
-    pub level: String,
-    pub format: LogFormat,
-    pub file: Option<String>,
-    pub access: AccessLogConfig,
-}
-
-impl Default for LoggingConfig {
-    fn default() -> Self {
-        Self {
-            level: "INFO".to_string(),
-            format: LogFormat::Pretty,
-            file: None,
-            access: AccessLogConfig::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct AccessLogConfig {
-    pub enabled: bool,
-    pub format: String,
-    pub file: Option<String>,
-}
-
-impl Default for AccessLogConfig {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            format: "[${time}] ${status} - ${method} ${path} ${latency}\n".to_string(),
-            file: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct BackendsConfig {
-    pub asset_studio: AssetStudioBackendConfig,
-    pub media: MediaBackendConfig,
-    pub image: ImageBackendConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct AssetStudioBackendConfig {
-    pub read_batch_size: usize,
-    pub image_format: Option<String>,
-    pub read_kinds: BTreeMap<String, String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct MediaBackendConfig {
-    pub backend: MediaBackend,
-    pub ffmpeg_path: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct ImageBackendConfig {
-    pub backend: ImageBackend,
-    pub png_compression: ImagePngCompression,
-    pub webp_lossless: bool,
-    pub jpeg_quality: u8,
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ImageBackend {
-    #[default]
-    Rust,
-}
-
-impl FromStr for ImageBackend {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "rust" => Ok(Self::Rust),
-            other => Err(ConfigError::InvalidValue {
-                field: "backends.image.backend".to_string(),
-                value: other.to_string(),
-                expected: "rust".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ImagePngCompression {
-    #[default]
-    Fast,
-    Default,
-    Best,
-}
-
-impl FromStr for ImagePngCompression {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "fast" => Ok(Self::Fast),
-            "default" => Ok(Self::Default),
-            "best" => Ok(Self::Best),
-            other => Err(ConfigError::InvalidValue {
-                field: "backends.image.png_compression".to_string(),
-                value: other.to_string(),
-                expected: "fast, default, or best".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ImageOutputFormat {
-    Png,
-    Jpg,
-    Webp,
-}
-
-impl FromStr for ImageOutputFormat {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "png" => Ok(Self::Png),
-            "jpg" | "jpeg" => Ok(Self::Jpg),
-            "webp" => Ok(Self::Webp),
-            other => Err(ConfigError::InvalidValue {
-                field: "export.images.formats".to_string(),
-                value: other.to_string(),
-                expected: "png, jpg, or webp".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum VideoOutputFormat {
-    M2v,
-    Mp4,
-}
-
-impl FromStr for VideoOutputFormat {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "m2v" => Ok(Self::M2v),
-            "mp4" => Ok(Self::Mp4),
-            other => Err(ConfigError::InvalidValue {
-                field: "export.video.formats".to_string(),
-                value: other.to_string(),
-                expected: "m2v or mp4".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum AudioOutputFormat {
-    Wav,
-    Flac,
-    Mp3,
-}
-
-impl FromStr for AudioOutputFormat {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "wav" => Ok(Self::Wav),
-            "flac" => Ok(Self::Flac),
-            "mp3" => Ok(Self::Mp3),
-            other => Err(ConfigError::InvalidValue {
-                field: "export.audio.formats".to_string(),
-                value: other.to_string(),
-                expected: "wav, flac, or mp3".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum MediaBackend {
-    Auto,
-    #[default]
-    Ffi,
-    Cli,
-}
-
-impl FromStr for MediaBackend {
-    type Err = ConfigError;
-
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        match value.trim().to_lowercase().as_str() {
-            "auto" => Ok(Self::Auto),
-            "ffi" => Ok(Self::Ffi),
-            "cli" => Ok(Self::Cli),
-            other => Err(ConfigError::InvalidValue {
-                field: "backends.media.backend".to_string(),
-                value: other.to_string(),
-                expected: "auto, ffi, or cli".to_string(),
-            }),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct ExecutionConfig {
-    pub timeout_seconds: u64,
-    pub allow_cancel: bool,
-    /// Optional persistent cache for deobfuscated AssetBundles. Cache entries
-    /// are stored below `<root>/<region>/<bundle path>` and reused by update
-    /// and prefetch runs before the network is consulted.
-    pub asset_bundle_cache_dir: Option<String>,
-    /// Soft process memory guard for bundle work.  When non-zero, bundle
-    /// downloads/native payloads acquire permits by estimated bundle size and
-    /// keep them until export/post-process finishes.
-    pub max_in_flight_bundle_bytes: usize,
-    /// How many successful downloads to accumulate before flushing the download
-    /// record to disk mid-run.  Set to `0` to disable mid-run flushing (record
-    /// is only written once at the end).  Mirrors Go's `batchSaveSize`.
-    pub batch_save_size: usize,
-    /// Maximum number of jobs whose heavy download/export pipeline may run at once. Extra jobs are
-    /// accepted (HTTP 202) but queue for a slot instead of all running concurrently. `0` = no limit.
-    pub max_concurrent_jobs: usize,
-    /// Cap on retained terminal (Completed/Failed/Cancelled) job snapshots kept in memory for the
-    /// jobs API. Oldest terminal jobs are evicted beyond this. `0` = keep all (unbounded).
-    pub retain_terminal_jobs: usize,
-    pub retry: RetryConfig,
-}
-
-impl Default for ExecutionConfig {
-    fn default() -> Self {
-        Self {
-            timeout_seconds: 300,
-            allow_cancel: true,
-            asset_bundle_cache_dir: None,
-            max_in_flight_bundle_bytes: 0,
-            batch_save_size: 50,
-            max_concurrent_jobs: 4,
-            retain_terminal_jobs: 256,
-            retry: RetryConfig::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct RetryConfig {
-    pub attempts: usize,
-    pub initial_backoff_ms: u64,
-    pub max_backoff_ms: u64,
-}
-
-impl Default for RetryConfig {
-    fn default() -> Self {
-        Self {
-            attempts: 4,
-            initial_backoff_ms: 1_000,
-            max_backoff_ms: 4_000,
-        }
-    }
-}
-
-impl Default for AssetStudioBackendConfig {
-    fn default() -> Self {
-        Self {
-            read_batch_size: 64,
-            image_format: None,
-            read_kinds: BTreeMap::new(),
-        }
-    }
-}
-
-impl Default for MediaBackendConfig {
-    fn default() -> Self {
-        Self {
-            backend: MediaBackend::Ffi,
-            ffmpeg_path: "ffmpeg".to_string(),
-        }
-    }
-}
-
-impl Default for ImageBackendConfig {
-    fn default() -> Self {
-        Self {
-            backend: ImageBackend::Rust,
-            png_compression: ImagePngCompression::Fast,
-            webp_lossless: true,
-            jpeg_quality: 95,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct ConcurrencyConfig {
-    pub auto_tune: bool,
-    pub download: usize,
-    pub upload: usize,
-    pub post_process: usize,
-    pub acb: usize,
-    pub usm: usize,
-    pub hca: usize,
-    /// Legacy aggregate media encode cap. New configs should prefer
-    /// audio_encode and video_encode.
-    pub media_encode: usize,
-    pub audio_encode: usize,
-    pub video_encode: usize,
-    pub images: usize,
-}
-
-impl Default for ConcurrencyConfig {
-    fn default() -> Self {
-        Self {
-            auto_tune: false,
-            download: 32,
-            upload: 4,
-            post_process: 16,
-            acb: 12,
-            usm: 6,
-            hca: 16,
-            media_encode: 12,
-            audio_encode: 12,
-            video_encode: 4,
-            images: 12,
-        }
-    }
-}
-
-impl ConcurrencyConfig {
-    pub fn effective(&self) -> Self {
-        if !self.auto_tune {
-            return self.clone();
-        }
-        self.effective_for_cpus(available_cpu_count())
-    }
-
-    pub fn effective_for_cpus(&self, cpus: usize) -> Self {
-        let cpu_budget = ResourcesConfig::default()
-            .cpu
-            .effective_budget_for_cpus(cpus.max(1));
-        self.effective_for_cpus_with_budget(cpus, cpu_budget)
-    }
-
-    /// Sizes each work pool for the host actually running the export.
-    ///
-    /// The configured value is a *floor*, not a target: each CPU-bound pool is
-    /// raised to a core-count-derived width and only then clamped by that
-    /// pool's cap. Before this scaled upward, the caps were the whole story, so
-    /// a config tuned on a 10-core host kept its literal numbers on a 64-core
-    /// one — measured on a 64-core EPYC 7B13, `audio_encode: 12` held the music
-    /// rule at exactly 12 busy cores for 290 s that widening finished in 86 s.
-    ///
-    /// The widths below come from sweeps on that host (JP corpus, all other
-    /// limits released), each stated as the point past which wall clock stopped
-    /// improving:
-    ///
-    /// - `post_process` — the budget. This one is easy to get wrong: a sweep
-    ///   with every *other* pool already at `2 x` the budget put its knee at
-    ///   `cpus / 2`, but that only held while post-processing was not the
-    ///   narrowest stage. Shipping `cpus / 2` here made it the narrowest, and
-    ///   the music rule pinned at exactly 31.5 busy cores against a 32-wide
-    ///   pool — `audio_encode` cannot be used above this number, because audio
-    ///   work happens inside a bundle's post-process slot. Widening this alone
-    ///   took that rule from 112 s to 74 s.
-    /// - `audio_encode` — knee at three quarters of the budget; every setting
-    ///   from there to `4 x` landed within run-to-run noise, so the budget
-    ///   itself is used.
-    /// - `video_encode` — kept at `cpus / 4` because each x264 instance is
-    ///   itself about six threads. Wider still measured slightly faster, but
-    ///   only by heavily oversubscribing the machine. Verified generous: the
-    ///   movie rule reached 54 of 64 cores at this width.
-    /// - `usm` — `cpus / 2`, not from a sweep of its own; it was never the
-    ///   binding pool in any run.
-    ///
-    /// Network pools (`download`, `upload`) are deliberately left alone: their
-    /// ceiling is the remote endpoint, not this host.
-    ///
-    /// On a narrow host every floor sits at or below the configured value, so
-    /// the result is byte-for-byte what the caps alone produced.
-    pub fn effective_for_cpus_with_budget(&self, cpus: usize, cpu_budget: usize) -> Self {
-        if !self.auto_tune {
-            return self.clone();
-        }
-        let cpus = cpus.max(1);
-        let cpu_budget = cpu_budget.max(1);
-        let cpu_oversubscribe = cpu_budget.saturating_mul(2).max(cpu_budget);
-        let half_cpus = cpus.div_ceil(2);
-        let quarter_cpus = cpus.div_ceil(4).max(1);
-        Self {
-            auto_tune: true,
-            download: self.download.min(cpus.saturating_mul(4).max(4)).max(1),
-            upload: self.upload.min(cpus.max(2)).max(1),
-            post_process: if self.post_process == 0 {
-                0
-            } else {
-                self.post_process
-                    .max(cpu_budget)
-                    .min(cpus.saturating_mul(2).max(2))
-                    .max(1)
-            },
-            acb: self.acb.max(cpu_budget).min(cpu_oversubscribe).max(1),
-            usm: self.usm.max(half_cpus).min(cpus.max(2)).max(1),
-            hca: self
-                .hca
-                .max(cpu_budget)
-                .min(cpus.saturating_mul(2).max(2))
-                .min(cpu_oversubscribe)
-                .max(1),
-            media_encode: self
-                .media_encode
-                .max(cpu_budget)
-                .min(cpu_oversubscribe)
-                .max(1),
-            audio_encode: self
-                .audio_encode
-                .max(cpu_budget)
-                .min(cpu_oversubscribe)
-                .max(1),
-            // Floor and cap coincide here, so the configured value never
-            // survives auto-tuning; leave `auto_tune` off to hand-hold x264 on
-            // a memory-constrained host.
-            video_encode: quarter_cpus,
-            images: self.images.max(cpu_budget).min(cpu_oversubscribe).max(1),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct ResourcesConfig {
-    pub cpu: CpuResourceConfig,
-    pub memory: MemoryResourceConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct CpuResourceConfig {
-    pub budget_auto: bool,
-    pub budget_ratio: f64,
-    pub reserved: usize,
-    pub throttle: CpuThrottleConfig,
-}
-
-impl Default for CpuResourceConfig {
-    fn default() -> Self {
-        Self {
-            budget_auto: true,
-            budget_ratio: 1.0,
-            reserved: 0,
-            throttle: CpuThrottleConfig::default(),
-        }
-    }
-}
-
-impl CpuResourceConfig {
-    pub fn effective_budget(&self) -> usize {
-        self.effective_budget_for_cpus(available_cpu_count())
-    }
-
-    pub fn effective_budget_for_cpus(&self, cpus: usize) -> usize {
-        let cpus = cpus.max(1);
-        if !self.budget_auto {
-            return cpus;
-        }
-        ((cpus as f64 * self.budget_ratio).floor() as usize)
-            .saturating_sub(self.reserved)
-            .max(1)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct CpuThrottleConfig {
-    pub enabled: bool,
-    pub sample_ms: u64,
-}
-
-impl Default for CpuThrottleConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            sample_ms: 250,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct MemoryResourceConfig {
-    pub max_in_flight_bundle_bytes: usize,
-}
-
-fn available_cpu_count() -> usize {
-    std::thread::available_parallelism()
-        .map(usize::from)
-        .unwrap_or(4)
-        .max(1)
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct StorageConfig {
-    pub providers: Vec<StorageProviderConfig>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct StorageProviderConfig {
-    pub name: Option<String>,
-    #[serde(alias = "kind")]
-    pub scheme: String,
-    pub root: Option<String>,
-    pub public_base_url: Option<String>,
-    #[serde(default, deserialize_with = "deserialize_storage_options")]
-    pub options: BTreeMap<String, String>,
-    pub endpoint: String,
-    pub tls: bool,
-    pub bucket: String,
-    pub prefix: Option<String>,
-    pub path_style: bool,
-    pub region: Option<String>,
-    pub public_read: bool,
-    pub access_key: Option<String>,
-    pub secret_key: Option<String>,
-}
-
-impl Default for StorageProviderConfig {
-    fn default() -> Self {
-        Self {
-            name: None,
-            scheme: "s3".to_string(),
-            root: None,
-            public_base_url: None,
-            options: BTreeMap::new(),
-            endpoint: String::new(),
-            tls: true,
-            bucket: String::new(),
-            prefix: None,
-            path_style: true,
-            region: None,
-            public_read: false,
-            access_key: None,
-            secret_key: None,
-        }
-    }
-}
-
-fn deserialize_storage_options<'de, D>(
-    deserializer: D,
-) -> Result<BTreeMap<String, String>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let raw = BTreeMap::<String, Value>::deserialize(deserializer)?;
-    raw.into_iter()
-        .map(|(key, value)| {
-            storage_option_value_to_string(value)
-                .map(|value| (key, value))
-                .map_err(de::Error::custom)
-        })
-        .collect()
-}
-
-fn storage_option_value_to_string(value: Value) -> Result<String, String> {
-    match value {
-        Value::Null => Ok(String::new()),
-        Value::Bool(value) => Ok(value.to_string()),
-        Value::Number(value) => Ok(value.to_string()),
-        Value::String(value) => Ok(value),
-        Value::Sequence(_) | Value::Mapping(_) | Value::Tagged(_) => {
-            Err("storage provider options must be scalar values".to_string())
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct GitSyncConfig {
-    pub chart_hashes: ChartHashConfig,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "lowercase")]
-pub enum GitSigningFormat {
-    #[default]
-    #[serde(alias = "openpgp")]
-    Gpg,
-    Ssh,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct ChartHashConfig {
-    pub enabled: bool,
-    pub repository_dir: Option<String>,
-    pub username: Option<String>,
-    pub email: Option<String>,
-    pub password: Option<String>,
-    pub sign_commits: bool,
-    pub signing_format: GitSigningFormat,
-    pub signing_key: Option<String>,
-    pub signing_program: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct RegionConfig {
-    pub enabled: bool,
-    pub provider: RegionProviderConfig,
-    pub crypto: CryptoConfig,
-    pub runtime: RegionRuntimeConfig,
-    pub paths: RegionPathsConfig,
-    pub filters: RegionFiltersConfig,
-    pub export: RegionExportConfig,
-    pub upload: RegionUploadConfig,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "snake_case")]
-pub enum RegionProviderConfig {
-    ColorfulPalette {
-        asset_info_url_template: String,
-        asset_bundle_url_template: String,
-        profile: String,
-        profile_hashes: BTreeMap<String, String>,
-        #[serde(default)]
-        required_cookies: bool,
-        #[serde(default)]
-        cookie_bootstrap_url: Option<String>,
-    },
-    Nuverse {
-        asset_version_url: String,
-        app_version: String,
-        asset_info_url_template: String,
-        asset_bundle_url_template: String,
-        #[serde(default)]
-        required_cookies: bool,
-        #[serde(default)]
-        cookie_bootstrap_url: Option<String>,
-    },
-}
-
-impl Default for RegionProviderConfig {
-    fn default() -> Self {
-        Self::ColorfulPalette {
-            asset_info_url_template: String::new(),
-            asset_bundle_url_template: String::new(),
-            profile: String::new(),
-            profile_hashes: BTreeMap::new(),
-            required_cookies: false,
-            cookie_bootstrap_url: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct CryptoConfig {
-    pub aes_key_hex: Option<String>,
-    pub aes_iv_hex: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct RegionRuntimeConfig {
-    pub unity_version: String,
-}
-
-impl Default for RegionRuntimeConfig {
-    fn default() -> Self {
-        Self {
-            unity_version: "2022.3.21f1".to_string(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct RegionPathsConfig {
-    pub asset_save_dir: Option<String>,
-    pub downloaded_asset_record_file: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct RegionFiltersConfig {
-    pub start_app: Vec<String>,
-    pub on_demand: Vec<String>,
-    pub skip: Vec<String>,
-    pub priority: Vec<String>,
-}
-
-pub const DEFAULT_ASSET_STUDIO_EXPORT_TYPES: &[&str] = &["all"];
-
-fn default_asset_studio_export_types() -> Vec<String> {
-    DEFAULT_ASSET_STUDIO_EXPORT_TYPES
-        .iter()
-        .map(|value| (*value).to_string())
-        .collect()
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct RegionExportConfig {
-    pub by_category: bool,
-    #[serde(default = "default_asset_studio_export_types")]
-    pub asset_studio_types: Vec<String>,
-    pub raw_bundles: Option<RawBundleExportConfig>,
-    pub haruki_3d: Haruki3dExportConfig,
-    pub usm: UsmExportConfig,
-    pub acb: AcbExportConfig,
-    pub hca: HcaExportConfig,
-    pub images: ImageExportConfig,
-    pub video: VideoExportConfig,
-    pub audio: AudioExportConfig,
-}
-
-impl Default for RegionExportConfig {
-    fn default() -> Self {
-        Self {
-            by_category: false,
-            asset_studio_types: default_asset_studio_export_types(),
-            raw_bundles: None,
-            haruki_3d: Haruki3dExportConfig::default(),
-            usm: UsmExportConfig::default(),
-            acb: AcbExportConfig::default(),
-            hca: HcaExportConfig::default(),
-            images: ImageExportConfig::default(),
-            video: VideoExportConfig::default(),
-            audio: AudioExportConfig::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct RawBundleExportConfig {
-    pub output_dir: Option<String>,
-    pub include: Vec<String>,
-    pub exclude: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct Haruki3dExportConfig {
-    pub enabled: bool,
-    pub exporter_path: String,
-    pub master_dir: String,
-    pub work_dir: String,
-    pub manifest_file: String,
-    pub staging_dir: String,
-    pub output_dir: String,
-    pub shared_content_store: String,
-    pub compiled_content_store: String,
-    pub process_concurrency: usize,
-    pub convert_model_textures: bool,
-    pub role_character3d_ids: Vec<i64>,
-    pub include: Vec<String>,
-    pub exclude: Vec<String>,
-    pub cleanup_work_dir_after_success: bool,
-    pub cleanup_work_dir_after_failure: bool,
-    pub cleanup_staging_after_success: bool,
-    pub cleanup_staging_after_failure: bool,
-}
-
-impl Default for Haruki3dExportConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            exporter_path: String::new(),
-            master_dir: String::new(),
-            work_dir: String::new(),
-            manifest_file: String::new(),
-            staging_dir: String::new(),
-            output_dir: String::new(),
-            shared_content_store: String::new(),
-            compiled_content_store: String::new(),
-            process_concurrency: 0,
-            convert_model_textures: false,
-            role_character3d_ids: Vec::new(),
-            include: Vec::new(),
-            exclude: Vec::new(),
-            cleanup_work_dir_after_success: true,
-            cleanup_work_dir_after_failure: true,
-            cleanup_staging_after_success: true,
-            cleanup_staging_after_failure: false,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct UsmExportConfig {
-    pub export: bool,
-    pub decode: bool,
-}
-
-impl Default for UsmExportConfig {
-    fn default() -> Self {
-        Self {
-            export: true,
-            decode: true,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct AcbExportConfig {
-    pub export: bool,
-    pub decode: bool,
-}
-
-impl Default for AcbExportConfig {
-    fn default() -> Self {
-        Self {
-            export: true,
-            decode: true,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default)]
-pub struct HcaExportConfig {
-    pub decode: bool,
-}
-
-impl Default for HcaExportConfig {
-    fn default() -> Self {
-        Self { decode: true }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct ImageExportConfig {
-    pub formats: Vec<ImageOutputFormat>,
-}
-
-impl Default for ImageExportConfig {
-    fn default() -> Self {
-        Self {
-            formats: vec![ImageOutputFormat::Png],
-        }
-    }
-}
-
-impl ImageExportConfig {
-    pub fn output_formats(&self) -> Vec<ImageOutputFormat> {
-        dedupe_image_formats(self.formats.clone())
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct VideoExportConfig {
-    pub formats: Vec<VideoOutputFormat>,
-    pub direct_mp4: bool,
-}
-
-impl Default for VideoExportConfig {
-    fn default() -> Self {
-        Self {
-            formats: vec![VideoOutputFormat::Mp4],
-            direct_mp4: false,
-        }
-    }
-}
-
-impl VideoExportConfig {
-    pub fn output_formats(&self) -> Vec<VideoOutputFormat> {
-        dedupe_video_formats(self.formats.clone())
-    }
-
-    pub fn writes_m2v(&self) -> bool {
-        self.output_formats().contains(&VideoOutputFormat::M2v)
-    }
-
-    pub fn writes_mp4(&self) -> bool {
-        self.output_formats().contains(&VideoOutputFormat::Mp4)
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(default, deny_unknown_fields)]
-pub struct AudioExportConfig {
-    pub formats: Vec<AudioOutputFormat>,
-}
-
-impl Default for AudioExportConfig {
-    fn default() -> Self {
-        Self {
-            formats: vec![AudioOutputFormat::Mp3],
-        }
-    }
-}
-
-impl AudioExportConfig {
-    pub fn output_formats(&self) -> Vec<AudioOutputFormat> {
-        dedupe_audio_formats(self.formats.clone())
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct RegionUploadConfig {
-    pub enabled: bool,
-    pub providers: Vec<String>,
-    pub public_read: UploadPublicReadConfig,
-    pub remove_local_after_upload: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-#[serde(default)]
-pub struct UploadPublicReadConfig {
-    pub include: Vec<String>,
-    pub exclude: Vec<String>,
-}
+//! Configuration: shape, loading, environment overrides, validation, tuning.
+//!
+//! This file is the facade. Every type is still reachable as
+//! `crate::core::config::Thing`; the submodules below hold the definitions.
+
+mod env;
+mod load;
+mod schema;
+mod tuning;
+mod validate;
+
+pub use schema::*;
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use std::io::Write;
+    use std::path::Path;
     use std::sync::{Mutex, MutexGuard, OnceLock};
 
     use tempfile::NamedTempFile;
+
+    use crate::core::errors::ConfigError;
+
+    use super::schema::default_asset_studio_export_types;
 
     fn env_lock() -> MutexGuard<'static, ()> {
         static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
